+++ packages/repos/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @happyvertical/repos
 
+## 0.89.1
+
+### Patch Changes
+
+- @happyvertical/graphql@0.89.1
+
+## 0.89.0
+
+### Patch Changes
+
+- @happyvertical/graphql@0.89.0
+
+## 0.88.2
+
+### Patch Changes
+
+- @happyvertical/graphql@0.88.2
+
+## 0.88.1
+
+### Patch Changes
+
+- @happyvertical/graphql@0.88.1
+
 ## 0.88.0
 
 ### Patch Changes