+++ packages/social/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @happyvertical/social
 
+## 0.89.1
+
+### Patch Changes
+
+- @happyvertical/logger@0.89.1
+- @happyvertical/utils@0.89.1
+
+## 0.89.0
+
+### Patch Changes
+
+- @happyvertical/logger@0.89.0
+- @happyvertical/utils@0.89.0
+
+## 0.88.2
+
+### Patch Changes
+
+- Updated dependencies [596e262]
+  - @happyvertical/utils@0.88.2
+  - @happyvertical/logger@0.88.2
+
+## 0.88.1
+
+### Patch Changes
+
+- @happyvertical/logger@0.88.1
+- @happyvertical/utils@0.88.1
+
 ## 0.88.0
 
 ### Patch Changes